layout/base/tests/bug977003_inner_6.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=977003
https://bugzilla.mozilla.org/show_bug.cgi?id=1073563
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bugs 977003, 1073563</title>
  <meta name="author" content="Maksim Lebedev" />
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style>
    #target, #listener { background: yellow; }
  </style>
  <script type="application/javascript">
    var target = undefined;
    var listener = undefined;
    var test_target = false;
    var test_move = false;
    var test_listener = false;
    var receive_lostpointercapture = false;

    function TargetDownHandler(event) {
      logger("Target receive event: " + event.type);
      logger("Send setPointerCapture to listener");
      listener.setPointerCapture(event.pointerId);
      logger("set/release was executed");
      test_target = true;
    }
    function ListenerHandler(event) {
      logger("Receive event on Listener: " + event.type);
      if("gotpointercapture" == event.type) {
        logger("Send releasePointerCapture from listener");
        listener.releasePointerCapture(event.pointerId);
      } else if(event.type == "lostpointercapture") {
        // Set/release pointer capture in the event listeners of got/lostpointercapture won't take effect immediately
        parent.is(receive_lostpointercapture, false, "Part 6: listener should receive only one lostpointercapture");
        if (!receive_lostpointercapture) {
          receive_lostpointercapture = true;
          logger("Send setPointerCapture to listener");
          listener.setPointerCapture(event.pointerId);
        }
      } else if(event.type == "pointermove") {
        test_move = true;
      } else {
        test_listener = true;
      }
    }
    function logger(message) {
      console.log(message);
      var log = document.getElementById('log');
      log.innerHTML = message + "<br>" + log.innerHTML;
    }

    function prepareTest() {
      parent.turnOnPointerEvents(executeTest);
    }
    function executeTest()
    {
      logger("executeTest");
      target = document.getElementById("target");
      listener = document.getElementById("listener");
      target.addEventListener("pointerdown",          TargetDownHandler,  false);
      listener.addEventListener("gotpointercapture",  ListenerHandler,    false);
      listener.addEventListener("pointerdown",        ListenerHandler,    false);
      listener.addEventListener("pointerover",        ListenerHandler,    false);
      listener.addEventListener("pointermove",        ListenerHandler,    false);
      listener.addEventListener("pointerup",          ListenerHandler,    false);
      listener.addEventListener("pointerout",         ListenerHandler,    false);
      listener.addEventListener("lostpointercapture", ListenerHandler,    false);
      var rect = target.getBoundingClientRect();
      synthesizePointer(target, rect.width/2, rect.height/2, {type: "pointerdown"});
      synthesizePointer(target, rect.width/2, rect.height/2, {type: "pointermove"});
      synthesizePointer(target, rect.width/2, rect.height/2, {type: "pointerup"});
      finishTest();
    }
    function finishTest() {
      parent.is(test_target,    true,   "Part 6: pointerdown event should be received by target");
      // PE level 2 defines that the pending pointer capture is processed when firing next pointer events.
      // In this test case, pointer capture release is processed when firing pointermove
      parent.is(test_move,      true,   "Part 6: gotpointercapture should be triggered by pointermove");
      parent.is(test_listener,  false,  "Part 6: no other pointerevents should be fired before gotpointercapture");
      logger("finishTest");
      parent.finishTest();
    }
  </script>
</head>
<body onload="prepareTest()">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=977003">Mozilla Bug 977003 Test 6</a>
  <br><a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1073563">Mozilla Bug 1073563</a>
  <p id="display"></p>
  <div id="content" style="display: none">
  </div>
  <div id="listener">div id=listener</div>
  <div id="target">div id=target</div>
  <pre id="log">
  </pre>
</body>
</html>
